Add Mojang profile fallback for offline mode servers
Players on offline mode servers have UUIDs that don't match a Mojang
profile, so the Mojang profile API returns empty. This causes the
getJSON() call in getProfile() to return an exception (a 204 error
code). The result is that the chat message doesn't get through to
Matrix, and isn't echoed back to Minecraft players.

Catch errors from getJSON() in getProfile(), and fall back to a default
profile with the player's UUID, display name, and the default steve
head.

To catch the error, we await the result of getJSON() from getProfile().
I'm new to typescript, but my understanding is this shouldn't impact
latency since getProfile() is itself async.