vibeCheck: Report errors similarly to other endpoints
Errors are reported from vibeCheck() in a non-standard way, using JSON
status="BAD" and with error containing the error type and message e.g.
from Errors.noTokenError. The other end points have the error type and
message at the top level of the JSON, and thats how the bukkit plugin
expects it too.

Therefore report errors from vibeCheck() in the same way that other
endpoints do so all errors can be handled in the same way by the plugin.