Provide structured events from a GET /events endpoint
To allow for events to be passed from Matrix->Minecraft that aren't
plain chat messages (for example kicks, bans and unbans), add a new
GET /events endpoint to return an array of structured events.

So far we define 3 events, for text messages, emotes, and announcements
(allowing them to potentially be formatted by the Minecraft plugin with
tooltips and command links).

For compatibility, the legacy GET /chat endpoint is kept working, by
converting the structured message events into plain messages and
discarding non-message events.