MatrixInterface: Minimally handle query.user
When an attempt is made to ban a non-existing Minecraft user, the
homeserver will query the appservice so it can create a new Matrix user.
This arrives in the form of a "query.user" event, but since we never
handle it, no response is sent to the homeserver, resulting in a large
delay before informing the appservice of the room ban event.

So minimally handle the event, calling the callback with false so that
the appservice responds with 404 and the homeserver can promptly deliver
the relevant events.