Unbridge by room ID, not bridge ID
Fix unbridge command to take a room ID rather than bridge ID.

The usage help says unbridging is done by room ID, but BridgeManager and
BridgeTable refer to "Bridge identifier" and match the database record
based on the id field rather than the room field. This results in the
rather confusing situation where a room can't be bridged because it
already is, but it can't be unbridged because it never was!

As well as more closely matching briding semantics, this also allows
unbridging to be performed even if the token has been lost for some
reason.