CmdManager: Fix privilege checks
Currently CmdManager only checks the power level of the user in the room
the command came from. This is not a reliable indicator of privilege for
a particular bridge, since the user may well have created that room
themselves, which allows anybody to unbridge any bridge or start a
bridge in a room they don't have power in.

Separately pass the command room and target room to checkPrivilege() so
it can check the power levels in the target room, but respond to the
user in the command room.

In the case of bridging, the bridge bot needs to be present in the room
for the power levels check to use the latest power levels state. We now
detect an absent bot (in the slightly unpleasant way of uncached
enumerating joined rooms - see the NOTE), and suggest the user invites
the bot in to the room first.

In the case of unbridging, the bridge bot is expected to be present in
the room, in which case the power levels check should be able to work.
If the bot has already been kicked, the power levels check can still
work but may use stale values from the time of the kick, which is
acceptable since the bridge may not function fully with the bot absent.
In the unlikely event this prevents a kicked bridge from being stopped
by a recently empowered user (or if the bot was never present), the bot
can always be re-invited to the room first.