MsgProcessor: Generate JSON formatted text from HTML
Generate intermediate JSON formatted text from org.matrix.custom.html
formatted_body in text and emote messages. The HTML parsing is done by
the node-html-parser library which is fast and lightweight. HTMLLayout
is then used to process paragraphs and whitespace in the resulting DOM,
which is then walked and converted to intermediate JSON.

Supported HTML tags include:
 - FONT / SPAN
   - data-mx-color/color -> type: "style", color: 24-bit number
   - data-mx-spoiler     -> type: "style", spoiler: true
 - B / STRONG            -> type: "style", bold: true
 - I / EM                -> type: "style", italic: true
 - U                     -> type: "style", underline: true
 - DEL / STRIKE          -> type: "style", strike: true
 - CODE                  -> type: "style", code: true
 - H1-H6                 -> type: "style", heading: number
 - UL / OL and LI        -> type: "block", block: "bullet", [n: number]
 - HR                    -> type: "horizontalRule"
 - BLOCKQUOTE            -> type: "block", block: "quote"
 - PRE			 -> type: "block", block: "preformat"
 - A href=matrix.to/@... -> type: "mention", user: ...
 - A href=...            -> type: "link", href=..., ...
 - IMG                   -> type: "img", [title: ...], [alt: ...], [src: media URL]

The BR, P & DIV tags are also supported, and are handled by HTMLLayout.

Also supported in both plain text and HTML formatted messages:
 - @room                 -> type: "mention", room=true

The following are not yet supported:
 - TABLE, TR, THEAD, TBODY, TH, TD, CAPTION
 - SUP/SUB - could maybe render as ^xx^ and _xx_?