Add HTMLLayout tests
Add some tests for the HTMLLayout class. Tests are fairly minimal and
directed and are not exhaustive, attempting to separately test each
piece of functionality of the HTMLLayout class, with a few edge cases
thrown in for good measure.

Each test parses some HTML to a DOM, processes the paragraphs &
whitespace in place using HTMLLayout (using a different paragraph
separator to normal to distinguish paragraphs from newlines within
paragraphs), and converts to a simplified JSON representation of the
HTML for comparison against the expected output.