Add MatrixInterface.formatMcMessage tests
Add some tests for the MatrixInterface.formatMcMessage() function which
converts a formatted Minecraft message into an HTML formatted Matrix
message. Tests are fairly minimal and directed and are not exhaustive,
attempting to separately test each formatting code that is handled, with
a few more complex edge cases thrown in for good measure.

A few tests are included for redundant formatting codes. Currently the
code is kept simple and produces correct but redundant HTML to match.
I've kept the tests using the expected output, but noted in comments
what the preferred output would be.

Each test creates a abbreviated Minecraft event, calls
formatMcMessage(), and expects particular values for the body and
formatted_text of the resulting Matrix event.