Add MsgProcessor.formatMxMessage tests
Add some tests for the MsgProcessor.formatMxMessage() function which
converts an HTML formatted Matrix message into the intermediate JSON
formatted text format. Tests are fairly minimal and directed and are not
exhaustive, attempting to separately test each piece of functionality
(including a number of colour formats, a bunch of HTML elements with
their attributes, and @room mentions).

Colour formats test MsgProcessor.htmlColorTo24bit() directly. The
remaining tests set up a Matrix event to pass to formatMxMessage(), and
expect a particular intermediate JSON for the 2nd result (the formatted
body).

The MatrixInterface and Config needed by the MsgProcessor class are
fairly simplistically mocked, allowing @room mentions to be roughly
tested for when the sender does and doesn't have permission to do so,
and for user mentions to be tested using a fake profile.